requirement.txt
pandas==1.4.3
streamlit==1.12.0
scikit-learn==1.1.2
sklearn-pandas==2.2.0
Unidecode==1.3.4
seaborn==0.11.2
matplotlib==3.5.3
palettable==3.3.0
plotly-express==0.4.1
plotly==5.10.0